separate leaflet configuration
